+++ NEWS.md
@@ -14,6 +14,7 @@
 * A step to remove simple zero-variance variables was added (`step_zv`).
 * A series of `tidy` methods were added for recipes and many (but not all) steps. 
 * In `bake.recipe`, the argument `newdata` is now without a default. 
+* A formula method was added for recipes to get a formula with the outcome(s) and predictors based on the trained recipe. 
 
 # recipes 0.1.0